+++ library/useAES.h
@@ -23,11 +23,11 @@ void aesDecrypt(uint8_t*, uint8_t*, uint8_t*);
     /* Performs a single AES decryption */
 void aesEncrypt(uint8_t*, uint8_t*, uint8_t*);
     /* Performs a single AES Encryption */
-void aesCBCdecrypt(uint8_t*, uint8_t*, uint8_t, uint8_t*, uint8_t*);
+void aesCBCdecrypt(uint8_t*, uint8_t*, uint16_t, uint8_t*, uint8_t*);
     /* Performs complete AES Decryption, CBC Mode */
-void aesCBCencrypt(uint8_t*, uint8_t*, uint8_t*, uint8_t*, uint8_t*);
+void aesCBCencrypt(uint8_t*, uint8_t*, uint16_t*, uint8_t*, uint8_t*);
     /* Performs complete AES Encryption, CBC Mode */
-void padding(uint8_t*, uint8_t*); 
+void padding(uint8_t*, uint16_t*); 
     /* Pads data until a multipe of 16 */
 
 #endif /* __USEAES_H__ */
